+++ protoc_plugin/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 16.0.0
+
+* Breaking change: Remove the '$checkItem' function from generated message classes and use the new method 'pc' on
+  'BuilderInfo' to add repeated composite fields.
+  Generated files require package:protobuf version 0.13.4 or newer.
+
 ## 15.0.3
 
 * Add test for frozen messages with unknown fields and update protobuf dependency to 0.13.1.